+++ frontend/public/locales/en/translation.json
@@ -2,11 +2,11 @@
   "index": {
     "betaTitle": "Beta Version",
     "betaMessage": "Please note that this is a beta version.",
-    "introduction": " The purpose of this application is to make your job easier when it comes to converting your docker-compose files into Kubernetes objects.",
+    "introduction": " The purpose of this application is to make your job easier when it comes to converting your docker-compose files into Kubernetes objects. This is the reason why we wanted to create an online tool to get your file back.",
     "steps": "You can convert your docker-compose files into three simple steps:",
-    "step1": "1. Click on the 'Choose File' button.",
+    "step1": "1. Click on the 'Choose File'.",
     "step2": "2. Select your docker-compose.yml file.",
-    "step3": "3. Press submit to convert your docker-compose into a Kubernetes file and get the result.",
+    "step3": "3. Submit",
     "disclaimer": "Disclaimer: We do not save any of your docker-compose files submitted."
   },
   "fileUpload": {
