+++ frontend/src/components/intro.jsx
@@ -3,18 +3,17 @@ import FileSyncImg from "../assets/images/file-sync.svg"
 import { makeStyles } from "@material-ui/core/styles"
 import { useTranslation } from "react-i18next"
 import Grid from "@material-ui/core/Grid"
-import Paper from "@material-ui/core/Paper"
-// import overlayImg from "../assets/illustrations/overlay.svg"
+import SampleFile from "../components/file-sample"
+import { Alert, AlertTitle } from "@material-ui/lab"
 
 const useStyles = makeStyles((theme) => ({
   root: {
     flexGrow: 1
-    // backgroundImage: `url(${overlayImg})`
   },
   img: {
     flex: "1",
-    width: "100%",
-    height: "100%"
+    width: "75%",
+    height: "70%"
   },
   startedBtn: {
     padding: "18px 36px",
@@ -41,7 +40,12 @@ const Intro = () => {
           <h2>What is Com2kube?</h2>
           <p>{t("index.introduction")}</p>
           <p>{t("index.steps")}</p>
-          <button className={classes.startedBtn}>Get Started</button>
+          <Alert severity="info">
+            <AlertTitle>{t("index.betaTitle")}</AlertTitle>
+            {t("index.betaMessage")}
+            <SampleFile />
+          </Alert>
+          <span>{t("index.disclaimer")}</span>
         </Grid>
         <Grid item xs={6}>
           <img src={FileSyncImg} alt="file sync" className={classes.img} />
