+++ frontend/src/components/stepper.jsx
@@ -1,13 +1,13 @@
 import React from "react"
 import { makeStyles } from "@material-ui/core/styles"
-import Stepper from "@material-ui/core/Stepper"
-import Step from "@material-ui/core/Step"
-import StepButton from "@material-ui/core/StepButton"
+import { useTranslation } from "react-i18next"
+import Breadcrumbs from "@material-ui/core/Breadcrumbs"
+import Typography from "@material-ui/core/Typography"
+import NavigateNextIcon from "@material-ui/icons/NavigateNext"
 
 const useStyles = makeStyles((theme) => ({
   root: {
-    width: "100%",
-    margin: theme.spacing(5, 2)
+    // width: "100%"
   },
   button: {
     marginRight: theme.spacing(1)
@@ -21,57 +21,18 @@ const useStyles = makeStyles((theme) => ({
   }
 }))
 
-function getSteps() {
-  return ["Select campaign settings", "Create an ad group", "Create an ad"]
-}
-
 export default function StepperSteps() {
   const classes = useStyles()
-  const [activeStep, setActiveStep] = React.useState(0)
-  const [completed, setCompleted] = React.useState({})
-  const steps = getSteps()
-
-  const totalSteps = () => {
-    return steps.length
-  }
-
-  const completedSteps = () => {
-    return Object.keys(completed).length
-  }
-
-  const isLastStep = () => {
-    return activeStep === totalSteps() - 1
-  }
-
-  const allStepsCompleted = () => {
-    return completedSteps() === totalSteps()
-  }
-
-  const handleNext = () => {
-    const newActiveStep =
-      isLastStep() && !allStepsCompleted()
-        ? // It's the last step, but not all steps have been completed,
-          // find the first step that has been completed
-          steps.findIndex((step, i) => !(i in completed))
-        : activeStep + 1
-    setActiveStep(newActiveStep)
-  }
-
-  const handleStep = (step) => () => {
-    setActiveStep(step)
-  }
+  const { t } = useTranslation()
 
   return (
-    <div className={classes.root}>
-      <Stepper nonLinear activeStep={activeStep}>
-        {steps.map((label, index) => (
-          <Step key={label}>
-            <StepButton onClick={handleStep(index)} completed={completed[index]}>
-              {label}
-            </StepButton>
-          </Step>
-        ))}
-      </Stepper>
-    </div>
+    <Breadcrumbs
+      separator={<NavigateNextIcon fontSize="large" />}
+      aria-label="breadcrumb"
+    >
+      <Typography color="textPrimary">{t("index.step1")}</Typography>
+      <Typography color="textPrimary">{t("index.step2")}</Typography>
+      <Typography color="textPrimary">{t("index.step3")}</Typography>
+    </Breadcrumbs>
   )
 }
