+++ frontend/src/themes/header/header.jsx
@@ -7,7 +7,6 @@ import Links from "./links/links"
 
 const useStyles = makeStyles(() => ({
   root: {
-    position: "fixed",
     flexGrow: 1,
     width: "100%",
     display: "flex",
